[cm] Split into two files
